Change CSS file
For reasons that are unclear, code blocks would display with large
amounts of white space at the beginning and end of the block unless
the `white-space` property was set to `normal` or `nowrap`. Doing this
meant that line indentation would not display.

The solution employed was to add a CSS rule that change the
`white-space` property for the `<div>` tags on individual lines.