Wrap code blocks in <code> tag
Previously code blocks would consist of <pre> and <div> tags only. However,
this causes issues in RSS readers which do not render the code blocks
correctly if they are not in this form. This commit ensures that code blocks
consist of both a <pre> and <code> tag and adjusts the CSS styling as
appropriate.